changes made for email

+++ src/Components/admin/layout/Header.jsx
@@ -3,14 +3,33 @@ import { Dropdown } from 'react-bootstrap'
 import { Link, useNavigate } from 'react-router-dom'
 import { useState } from "react";
 import { getTokenData, getInitials } from "../../../Utils";
+import { confirm_swal_with_text,success_swal_toast } from '../../../SwalServices';
 const Sidebar = React.lazy(() => import('./Sidebar'));
 
 function Header() {
     const navigate = useNavigate();
-    const logout = () => {
-        localStorage.clear();
-        navigate('/')
+    // const logout = () => {
+    //     localStorage.clear();
+    //     navigate('/')
+    // }
+      const logout = () => {
+  confirm_swal_with_text(async (resolve) => {
+    try {
+      // Clear local storage or token
+      localStorage.clear();
+
+      // Optional: show success message
+      success_swal_toast("You have been logged out!");
+
+      // Navigate to homepage or login
+      navigate("/");
+
+      resolve(true);
+    } catch (err) {
+      console.error("Logout error:", err);
     }
+  }, "Are you sure you want to logout?");
+}
     const [fullName, setFullName] = useState('');
 
     useEffect(() => {
